update through chapter 8
