+++ content/resources/chapter-8/index.ts
@@ -1,7 +1,17 @@
-import * as BuildingBlocksResources from './building-blocks'
+import * as BuildingBlocksResourcesThree from './building-blocks-3'
+import * as BuildingBlocksResourcesFour from './building-blocks-4'
+import * as BuildingBlocksResourcesFive from './building-blocks-5'
+import * as BuildingBlocksResourcesSix from './building-blocks-6'
+import * as BuildingBlocksResourcesSeven from './building-blocks-7'
+import * as BuildingBlocksResourcesEight from './building-blocks-8'
 
 const Chapter8 = {
-  'building-blocks': BuildingBlocksResources,
+  'building-blocks-3': BuildingBlocksResourcesThree,
+  'building-blocks-4': BuildingBlocksResourcesFour,
+  'building-blocks-5': BuildingBlocksResourcesFive,
+  'building-blocks-6': BuildingBlocksResourcesSix,
+  'building-blocks-7': BuildingBlocksResourcesSeven,
+  'building-blocks-8': BuildingBlocksResourcesEight,
 }
 
 export default Chapter8
